requirements.txt
pandas
pytest
streamlit
spotipy
python-dotenv
groq
